Split out multi-compiler functionality
