+++ lib/MultiCompiler.js
@@ -4,23 +4,8 @@
 */
 var Tapable = require("tapable");
 var async = require("async");
-var Stats = require("./Stats");
-
-function MultiWatching(watchings) {
-	this.watchings = watchings;
-}
-
-MultiWatching.prototype.invalidate = function() {
-	this.watchings.forEach(function(watching) {
-		watching.invalidate();
-	});
-};
-
-MultiWatching.prototype.close = function(callback) {
-	async.forEach(this.watchings, function(watching, callback) {
-		watching.close(callback);
-	}, callback);
-};
+var MultiWatching = require("./MultiWatching");
+var MultiStats = require("./MultiStats");
 
 function MultiCompiler(compilers) {
 	Tapable.call(this);
@@ -190,55 +175,3 @@ MultiCompiler.prototype.purgeInputFileSystem = function() {
 			compiler.inputFileSystem.purge();
 	});
 };
-
-function MultiStats(stats) {
-	this.stats = stats;
-	this.hash = stats.map(function(stat) {
-		return stat.hash;
-	}).join("");
-}
-
-MultiStats.prototype.hasErrors = function() {
-	return this.stats.map(function(stat) {
-		return stat.hasErrors();
-	}).reduce(function(a, b) {
-		return a || b;
-	}, false);
-};
-
-MultiStats.prototype.hasWarnings = function() {
-	return this.stats.map(function(stat) {
-		return stat.hasWarnings();
-	}).reduce(function(a, b) {
-		return a || b;
-	}, false);
-};
-
-MultiStats.prototype.toJson = function(options, forToString) {
-	var jsons = this.stats.map(function(stat) {
-		var obj = stat.toJson(options, forToString);
-		obj.name = stat.compilation && stat.compilation.name;
-		return obj;
-	});
-	var obj = {
-		errors: jsons.reduce(function(arr, j) {
-			return arr.concat(j.errors.map(function(msg) {
-				return "(" + j.name + ") " + msg;
-			}));
-		}, []),
-		warnings: jsons.reduce(function(arr, j) {
-			return arr.concat(j.warnings.map(function(msg) {
-				return "(" + j.name + ") " + msg;
-			}));
-		}, [])
-	};
-	if(!options || options.version !== false)
-		obj.version = require("../package.json").version;
-	if(!options || options.hash !== false)
-		obj.hash = this.hash;
-	if(!options || options.children !== false)
-		obj.children = jsons;
-	return obj;
-};
-
-MultiStats.prototype.toString = Stats.prototype.toString;
